+++ R/defunct.R
@@ -17,3 +17,10 @@ display <- function(...){
   .Defunct("compose")
 }
 
+#' @export
+#' @rdname flextable-defunct
+#' @details `ph_with_flextable()` is replaced by `officer::ph_with`.
+ph_with_flextable <- function( ... ){
+  .Defunct("officer::ph_with")
+}
+